add runtime comp. make department budget int

+++ BangazonEFTest/Models/Department.cs
@@ -15,7 +15,7 @@ namespace BangazonEFTest.Models
         public string Name { get; set; }
 
         [Required]
-        public double Budget { get; set; }
+        public int Budget { get; set; }
 
         public List<Employee> Employees { get; set; }
     }
